completely moved currentLocation functionality to the component
